+++ NEWS.md
@@ -1,10 +1,11 @@
 
-# rqdatatable 1.1.2 2018/12/10
+# rqdatatable 1.1.2 2018/12/13
 
  * Allow more control of ordering in extend.
  * Relax column production check.
  * Add rq_ufn().
  * More of force parent.frame forcing.
+ * Add row limit to order.
  * Add order_expr.
  * Add power test.